backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
motor==3.3.2
beanie==1.25.0
pydantic[email]==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cloudinary==1.39.0
httpx==0.27.0
